requirements.txt
pandas==0.23.4
numpy==1.14.5
scipy==0.19.1
nltk==3.3
gensim==3.5.0
textblob==0.15.2
googletrans==2.4.0
python_dateutil==2.8.0
scikit_learn==0.20.3
PyYAML==3.13
